+++ lets/decode/base64.py
@@ -1,4 +1,4 @@
-from lets.__module__ import Module
+from lets.__module__ import Module, TestCase
 import base64
 
 
@@ -12,13 +12,7 @@ class Base64(Module):
             yield base64.b64decode(data)
 
 
-from unittest import TestCase
-
 class Base64Tests(TestCase):
-
-    def setUp(self):
-        import importlib
-        self.module = importlib.import_module(__name__)
 
     def test_all_bytes(self):
         test = b"AAECAwQFBgcICQoLDA0ODxAREhMUFRYXGBkaGxwdHh8gISIjJCUmJygpKissLS4vMDEyMzQ1Njc4OTo7PD0+P0BBQkNERUZHSElKS0xNTk9QUVJTVFVWV1hZWltcXV5fYGFiY2RlZmdoaWprbG1ub3BxcnN0dXZ3eHl6e3x9fn+AgYKDhIWGh4iJiouMjY6PkJGSk5SVlpeYmZqbnJ2en6ChoqOkpaanqKmqq6ytrq+wsbKztLW2t7i5uru8vb6/wMHCw8TFxsfIycrLzM3Oz9DR0tPU1dbX2Nna29zd3t/g4eLj5OXm5+jp6uvs7e7v8PHy8/T19vf4+fr7/P3+/w=="